+++ userprovided/parameters.py
@@ -144,6 +144,20 @@ def int_in_range(parameter_name: str,
                                 fallback_value))
 
 
+def is_port(port_number: int) -> bool:
+    """Check if the number provided is valid as a TCP/UDP port
+       (i.e an integer in the range from 0 to 65535)."""
+
+    if not isinstance(port_number, int):
+        raise ValueError('Port has to be an integer.')
+
+    if 0 < port_number < 65536:
+        logging.debug('Port within range')
+        return True
+    logging.error('Port not within valid range from 0 to 65535')
+    return False
+
+
 def string_in_range(string_to_check: str,
                     minimum_length: int,
                     maximum_lenght: int,